extract formatBytes
